requirements.txt
langchain==0.2.4
streamlit==1.35.0
google-ai-generativelanguage==0.6.4
google-generativeai==0.6.0
plotly==5.22.0